#50 - Implement redirection for versions
For example, now
http://localhost:8080/user/en/4.7/advanced-configuration/change-logging
will redirect to:
http://localhost:8080/user/en/latest/initial-set-up/change-logging/

This example is doing two redirects -- one for the version `4.7`
redirecting to `latest` and another for the path.

Version redirects are stored in the book's yaml config in this repo.

Path redirects are store in the `redirects.txt` file within the book,
which was already working before this commit.